refactor(reservationController): 삭제할 id에 해당하는 예약을 찾을 수 없을 경우 204를 응답하도록 변경

+++ src/main/java/roomescape/controller/ReservationController.java
@@ -42,6 +42,9 @@ public class ReservationController {
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> delete(@PathVariable Long id) {
+        if (!reservationDao.isExist(id)) {
+            return ResponseEntity.noContent().build();
+        }
         reservationDao.delete(id);
         return ResponseEntity.ok().build();
     }
